+++ compiler-core/src/erlang/tests/bit_arrays.rs
@@ -209,3 +209,62 @@ fn identity(x) {
 "
     );
 }
+
+#[test]
+fn operator_in_pattern_size() {
+    assert_erl!(
+        "
+pub fn main() {
+  let assert <<len, payload:size(len * 8 - 8)>> = <<>>
+}
+"
+    );
+}
+
+#[test]
+fn operator_in_pattern_size2() {
+    assert_erl!(
+        "
+pub fn main() {
+  let assert <<len, payload:size(len / 8 - 1)>> = <<>>
+}
+"
+    );
+}
+
+#[test]
+fn operator_in_pattern_size3() {
+    assert_erl!(
+        "
+pub fn main() {
+  let additional = 10
+  let assert <<len, payload:size(len + additional * 8)>> = <<>>
+}
+"
+    );
+}
+
+#[test]
+fn block_in_pattern_size() {
+    assert_erl!(
+        "
+pub fn main() {
+  let assert <<len, payload:size({ len - 1 } * 8)>> = <<>>
+}
+"
+    );
+}
+
+#[test]
+fn non_byte_aligned_size_calculation() {
+    assert_erl!(
+        "
+pub fn main() {
+  case <<>> {
+    <<a:1, b:3, c:size(b - 2)>> -> c + b
+    _ -> 1
+  }
+}
+"
+    );
+}